+++ src/main/java/com/atlauncher/data/Constants.java
@@ -2,7 +2,7 @@ package com.atlauncher.data;
 
 
 public class Constants {
-    public static final String launcherName = "OpenLauncher";
+    public static final String LAUNCHER_NAME = "OpenLauncher";
     public static LauncherVersion VERSION = new LauncherVersion(1, 1, 1, 0, 0);
     public static final String API_BASE_URL = "";
     public static final String PASTE_CHECK_URL = "http://rushmead.playat.ch/stikket/";