+++ src/main/java/com/atlauncher/data/Instance.java
@@ -468,7 +468,15 @@ public class Instance implements Cloneable {
         if (this.ignoredUpdates == null) {
             this.ignoredUpdates = new ArrayList<String>();
         }
-        String version = getLatestVersion();
+
+        String version;
+
+        if (this.isDev) {
+            version = getLatestDevHash();
+        } else {
+            version = getLatestVersion();
+        }
+
         if (!hasUpdateBeenIgnored(version)) {
             this.ignoredUpdates.add(version);
             App.settings.saveInstances();
@@ -997,7 +1005,13 @@ public class Instance implements Cloneable {
      * there is no versions of the Pack
      */
     public String getLatestVersion() {
-        return (this.realPack != null ? this.realPack.getLatestVersion().getVersion() : null);
+        return (this.realPack != null ? (this.realPack.getLatestVersion() == null ? null : this.realPack
+                .getLatestVersion().getVersion()) : null);
+    }
+
+    public String getLatestDevHash() {
+        return (this.realPack != null ? (this.realPack.getLatestDevVersion() == null ? null : this.realPack
+                .getLatestDevVersion().getHash()) : null);
     }
 
     /**