+++ src/main/java/com/atlauncher/gui/LauncherConsole.java
@@ -54,7 +54,7 @@ public class LauncherConsole extends JFrame implements RelocalizationListener {
     private JMenuItem copy;
 
     public LauncherConsole() {
-        this.setTitle(Constants.launcherName + " Console " + Constants.VERSION);
+        this.setTitle(Constants.LAUNCHER_NAME + " Console " + Constants.VERSION);
         this.setDefaultCloseOperation(JFrame.HIDE_ON_CLOSE);
         this.setIconImage(Utils.getImage("/assets/image/Icon.png"));
         this.setMinimumSize(new Dimension(600, 400));