+++ src/main/java/com/atlauncher/gui/theme/Theme.java
@@ -19,6 +19,7 @@
 package com.atlauncher.gui.theme;
 
 import com.atlauncher.LogManager;
+import com.atlauncher.data.Constants;
 import com.atlauncher.utils.Resources;
 import com.atlauncher.utils.Utils;
 
@@ -27,14 +28,12 @@ import javax.swing.*;
 import java.awt.*;
 
 public final class Theme {
-    public static final Theme DEFAULT_THEME = new Theme("UntitledLauncher", "Modmuss50", true, new Color(255, 255,
+    public static final Theme DEFAULT_THEME = new Theme(Constants.LAUNCHER_NAME, "Modmuss50", true, new Color(255, 255,
             255), new Color(40, 45, 50), new Color(231, 231, 231), new Color(0, 136, 204), new Color(40, 45, 50),
             new Color(80, 170, 107), new Color(255, 255, 255), new Color(40, 25, 50), new Color(255, 255, 255),
             new Color(255, 255, 255), new Color(255, 0, 0), new Color(0, 0, 0), new Color(0, 0, 0),
             new Color(145, 0, 255), new Color(255, 0, 0), new Color(0, 0, 0), "SansSerif", "SansSerif",
             "Oswald-Regular", "SansSerif");
-
-
     // Meta
     private final String name;
     private final String author;