+++ src/main/java/modmuss50/mods/CustomPacks/UploadPackCode.java
@@ -26,7 +26,7 @@ public class UploadPackCode  implements Callable<String>{
         public String call() throws Exception {
         String log = code;
         String urlParameters = "";
-        urlParameters += "title=" + URLEncoder.encode(Constants.launcherName + " - Log", "ISO-8859-1") + "&";
+        urlParameters += "title=" + URLEncoder.encode(Constants.LAUNCHER_NAME + " - Log", "ISO-8859-1") + "&";
         urlParameters += "language=" + URLEncoder.encode("text", "ISO-8859-1") + "&";
         urlParameters += "private=" + URLEncoder.encode("1", "ISO-8859-1") + "&";
         urlParameters += "text=" + URLEncoder.encode(log, "ISO-8859-1") + "&";